Move shouldRun to run
